:technologist: unify firebase mock definition
