--- test/infrastructure/auth/email_and_password_auth_repository_test.mocks.dart
@@ -1,317 +1,0 @@
-// Mocks generated by Mockito 5.3.0 from annotations
-// in virtualpilgrimage/test/infrastructure/auth/email_and_password_auth_repository_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
-
-import 'package:firebase_auth/firebase_auth.dart' as _i4;
-import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart'
-    as _i3;
-import 'package:firebase_core/firebase_core.dart' as _i2;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeFirebaseApp_0 extends _i1.SmartFake implements _i2.FirebaseApp {
-  _FakeFirebaseApp_0(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeActionCodeInfo_1 extends _i1.SmartFake
-    implements _i3.ActionCodeInfo {
-  _FakeActionCodeInfo_1(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeUserCredential_2 extends _i1.SmartFake
-    implements _i4.UserCredential {
-  _FakeUserCredential_2(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeConfirmationResult_3 extends _i1.SmartFake
-    implements _i4.ConfirmationResult {
-  _FakeConfirmationResult_3(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-/// A class which mocks [FirebaseAuth].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockFirebaseAuth extends _i1.Mock implements _i4.FirebaseAuth {
-  MockFirebaseAuth() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.FirebaseApp get app => (super.noSuchMethod(Invocation.getter(#app),
-          returnValue: _FakeFirebaseApp_0(this, Invocation.getter(#app)))
-      as _i2.FirebaseApp);
-  @override
-  set app(_i2.FirebaseApp? _app) =>
-      super.noSuchMethod(Invocation.setter(#app, _app),
-          returnValueForMissingStub: null);
-  @override
-  set tenantId(String? tenantId) =>
-      super.noSuchMethod(Invocation.setter(#tenantId, tenantId),
-          returnValueForMissingStub: null);
-  @override
-  Map<dynamic, dynamic> get pluginConstants =>
-      (super.noSuchMethod(Invocation.getter(#pluginConstants),
-          returnValue: <dynamic, dynamic>{}) as Map<dynamic, dynamic>);
-  @override
-  _i5.Future<void> useEmulator(String? origin) => (super.noSuchMethod(
-      Invocation.method(#useEmulator, [origin]),
-      returnValue: _i5.Future<void>.value(),
-      returnValueForMissingStub: _i5.Future<void>.value()) as _i5.Future<void>);
-  @override
-  _i5.Future<void> useAuthEmulator(String? host, int? port) =>
-      (super.noSuchMethod(Invocation.method(#useAuthEmulator, [host, port]),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<void> applyActionCode(String? code) => (super.noSuchMethod(
-      Invocation.method(#applyActionCode, [code]),
-      returnValue: _i5.Future<void>.value(),
-      returnValueForMissingStub: _i5.Future<void>.value()) as _i5.Future<void>);
-  @override
-  _i5.Future<_i3.ActionCodeInfo> checkActionCode(String? code) =>
-      (super.noSuchMethod(Invocation.method(#checkActionCode, [code]),
-              returnValue: _i5.Future<_i3.ActionCodeInfo>.value(
-                  _FakeActionCodeInfo_1(
-                      this, Invocation.method(#checkActionCode, [code]))))
-          as _i5.Future<_i3.ActionCodeInfo>);
-  @override
-  _i5.Future<void> confirmPasswordReset({String? code, String? newPassword}) =>
-      (super.noSuchMethod(
-              Invocation.method(#confirmPasswordReset, [],
-                  {#code: code, #newPassword: newPassword}),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<_i4.UserCredential> createUserWithEmailAndPassword(
-          {String? email, String? password}) =>
-      (super
-          .noSuchMethod(Invocation.method(#createUserWithEmailAndPassword, [], {#email: email, #password: password}),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#createUserWithEmailAndPassword, [], {
-                        #email: email,
-                        #password: password
-                      })))) as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<List<String>> fetchSignInMethodsForEmail(String? email) => (super
-          .noSuchMethod(Invocation.method(#fetchSignInMethodsForEmail, [email]),
-              returnValue: _i5.Future<List<String>>.value(<String>[]))
-      as _i5.Future<List<String>>);
-  @override
-  _i5.Future<_i4.UserCredential> getRedirectResult() =>
-      (super.noSuchMethod(Invocation.method(#getRedirectResult, []),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#getRedirectResult, []))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  bool isSignInWithEmailLink(String? emailLink) => (super.noSuchMethod(
-      Invocation.method(#isSignInWithEmailLink, [emailLink]),
-      returnValue: false) as bool);
-  @override
-  _i5.Stream<_i4.User?> authStateChanges() =>
-      (super.noSuchMethod(Invocation.method(#authStateChanges, []),
-          returnValue: _i5.Stream<_i4.User?>.empty()) as _i5.Stream<_i4.User?>);
-  @override
-  _i5.Stream<_i4.User?> idTokenChanges() =>
-      (super.noSuchMethod(Invocation.method(#idTokenChanges, []),
-          returnValue: _i5.Stream<_i4.User?>.empty()) as _i5.Stream<_i4.User?>);
-  @override
-  _i5.Stream<_i4.User?> userChanges() =>
-      (super.noSuchMethod(Invocation.method(#userChanges, []),
-          returnValue: _i5.Stream<_i4.User?>.empty()) as _i5.Stream<_i4.User?>);
-  @override
-  _i5.Future<void> sendPasswordResetEmail(
-          {String? email, _i3.ActionCodeSettings? actionCodeSettings}) =>
-      (super.noSuchMethod(
-              Invocation.method(#sendPasswordResetEmail, [],
-                  {#email: email, #actionCodeSettings: actionCodeSettings}),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<void> sendSignInLinkToEmail(
-          {String? email, _i3.ActionCodeSettings? actionCodeSettings}) =>
-      (super.noSuchMethod(
-              Invocation.method(#sendSignInLinkToEmail, [],
-                  {#email: email, #actionCodeSettings: actionCodeSettings}),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<void> setLanguageCode(String? languageCode) => (super.noSuchMethod(
-      Invocation.method(#setLanguageCode, [languageCode]),
-      returnValue: _i5.Future<void>.value(),
-      returnValueForMissingStub: _i5.Future<void>.value()) as _i5.Future<void>);
-  @override
-  _i5.Future<void> setSettings(
-          {bool? appVerificationDisabledForTesting,
-          String? userAccessGroup,
-          String? phoneNumber,
-          String? smsCode,
-          bool? forceRecaptchaFlow}) =>
-      (super.noSuchMethod(
-              Invocation.method(#setSettings, [], {
-                #appVerificationDisabledForTesting:
-                    appVerificationDisabledForTesting,
-                #userAccessGroup: userAccessGroup,
-                #phoneNumber: phoneNumber,
-                #smsCode: smsCode,
-                #forceRecaptchaFlow: forceRecaptchaFlow
-              }),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<void> setPersistence(_i3.Persistence? persistence) =>
-      (super.noSuchMethod(Invocation.method(#setPersistence, [persistence]),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<_i4.UserCredential> signInAnonymously() =>
-      (super.noSuchMethod(Invocation.method(#signInAnonymously, []),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#signInAnonymously, []))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithCredential(
-          _i3.AuthCredential? credential) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithCredential, [credential]),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithCredential, [credential]))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithCustomToken(String? token) =>
-      (super.noSuchMethod(Invocation.method(#signInWithCustomToken, [token]),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithCustomToken, [token]))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithEmailAndPassword(
-          {String? email, String? password}) =>
-      (super
-          .noSuchMethod(Invocation.method(#signInWithEmailAndPassword, [], {#email: email, #password: password}),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#signInWithEmailAndPassword, [], {
-                        #email: email,
-                        #password: password
-                      })))) as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithEmailLink(
-          {String? email, String? emailLink}) =>
-      (super
-          .noSuchMethod(Invocation.method(#signInWithEmailLink, [], {#email: email, #emailLink: emailLink}),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#signInWithEmailLink, [], {
-                        #email: email,
-                        #emailLink: emailLink
-                      })))) as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithAuthProvider(
-          _i3.AuthProvider? provider) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithAuthProvider, [provider]),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithAuthProvider, [provider]))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<_i4.ConfirmationResult> signInWithPhoneNumber(String? phoneNumber,
-          [_i4.RecaptchaVerifier? verifier]) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithPhoneNumber, [phoneNumber, verifier]),
-              returnValue: _i5.Future<_i4.ConfirmationResult>.value(
-                  _FakeConfirmationResult_3(
-                      this,
-                      Invocation.method(
-                          #signInWithPhoneNumber, [phoneNumber, verifier]))))
-          as _i5.Future<_i4.ConfirmationResult>);
-  @override
-  _i5.Future<_i4.UserCredential> signInWithPopup(_i3.AuthProvider? provider) =>
-      (super.noSuchMethod(Invocation.method(#signInWithPopup, [provider]),
-              returnValue: _i5.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#signInWithPopup, [provider]))))
-          as _i5.Future<_i4.UserCredential>);
-  @override
-  _i5.Future<void> signInWithRedirect(_i3.AuthProvider? provider) =>
-      (super.noSuchMethod(Invocation.method(#signInWithRedirect, [provider]),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-  @override
-  _i5.Future<void> signOut() => (super.noSuchMethod(
-      Invocation.method(#signOut, []),
-      returnValue: _i5.Future<void>.value(),
-      returnValueForMissingStub: _i5.Future<void>.value()) as _i5.Future<void>);
-  @override
-  _i5.Future<String> verifyPasswordResetCode(String? code) =>
-      (super.noSuchMethod(Invocation.method(#verifyPasswordResetCode, [code]),
-          returnValue: _i5.Future<String>.value('')) as _i5.Future<String>);
-  @override
-  _i5.Future<void> verifyPhoneNumber(
-          {String? phoneNumber,
-          _i3.PhoneMultiFactorInfo? multiFactorInfo,
-          _i3.PhoneVerificationCompleted? verificationCompleted,
-          _i3.PhoneVerificationFailed? verificationFailed,
-          _i3.PhoneCodeSent? codeSent,
-          _i3.PhoneCodeAutoRetrievalTimeout? codeAutoRetrievalTimeout,
-          String? autoRetrievedSmsCodeForTesting,
-          Duration? timeout = const Duration(seconds: 30),
-          int? forceResendingToken,
-          _i3.MultiFactorSession? multiFactorSession}) =>
-      (super.noSuchMethod(
-              Invocation.method(#verifyPhoneNumber, [], {
-                #phoneNumber: phoneNumber,
-                #multiFactorInfo: multiFactorInfo,
-                #verificationCompleted: verificationCompleted,
-                #verificationFailed: verificationFailed,
-                #codeSent: codeSent,
-                #codeAutoRetrievalTimeout: codeAutoRetrievalTimeout,
-                #autoRetrievedSmsCodeForTesting: autoRetrievedSmsCodeForTesting,
-                #timeout: timeout,
-                #forceResendingToken: forceResendingToken,
-                #multiFactorSession: multiFactorSession
-              }),
-              returnValue: _i5.Future<void>.value(),
-              returnValueForMissingStub: _i5.Future<void>.value())
-          as _i5.Future<void>);
-}
-
-/// A class which mocks [UserCredential].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockUserCredential extends _i1.Mock implements _i4.UserCredential {
-  MockUserCredential() {
-    _i1.throwOnMissingStub(this);
-  }
-}
