--- test/infrastructure/auth/google_auth_repository_test.mocks.dart
@@ -1,429 +1,0 @@
-// Mocks generated by Mockito 5.3.0 from annotations
-// in virtualpilgrimage/test/infrastructure/auth/google_auth_repository_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i6;
-
-import 'package:firebase_auth/firebase_auth.dart' as _i4;
-import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart'
-    as _i3;
-import 'package:firebase_core/firebase_core.dart' as _i2;
-import 'package:google_sign_in/google_sign_in.dart' as _i5;
-import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart'
-    as _i7;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeFirebaseApp_0 extends _i1.SmartFake implements _i2.FirebaseApp {
-  _FakeFirebaseApp_0(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeActionCodeInfo_1 extends _i1.SmartFake
-    implements _i3.ActionCodeInfo {
-  _FakeActionCodeInfo_1(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeUserCredential_2 extends _i1.SmartFake
-    implements _i4.UserCredential {
-  _FakeUserCredential_2(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeConfirmationResult_3 extends _i1.SmartFake
-    implements _i4.ConfirmationResult {
-  _FakeConfirmationResult_3(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeGoogleSignInAuthentication_4 extends _i1.SmartFake
-    implements _i5.GoogleSignInAuthentication {
-  _FakeGoogleSignInAuthentication_4(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-/// A class which mocks [FirebaseAuth].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockFirebaseAuth extends _i1.Mock implements _i4.FirebaseAuth {
-  MockFirebaseAuth() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.FirebaseApp get app => (super.noSuchMethod(Invocation.getter(#app),
-          returnValue: _FakeFirebaseApp_0(this, Invocation.getter(#app)))
-      as _i2.FirebaseApp);
-  @override
-  set app(_i2.FirebaseApp? _app) =>
-      super.noSuchMethod(Invocation.setter(#app, _app),
-          returnValueForMissingStub: null);
-  @override
-  set tenantId(String? tenantId) =>
-      super.noSuchMethod(Invocation.setter(#tenantId, tenantId),
-          returnValueForMissingStub: null);
-  @override
-  Map<dynamic, dynamic> get pluginConstants =>
-      (super.noSuchMethod(Invocation.getter(#pluginConstants),
-          returnValue: <dynamic, dynamic>{}) as Map<dynamic, dynamic>);
-  @override
-  _i6.Future<void> useEmulator(String? origin) => (super.noSuchMethod(
-      Invocation.method(#useEmulator, [origin]),
-      returnValue: _i6.Future<void>.value(),
-      returnValueForMissingStub: _i6.Future<void>.value()) as _i6.Future<void>);
-  @override
-  _i6.Future<void> useAuthEmulator(String? host, int? port) =>
-      (super.noSuchMethod(Invocation.method(#useAuthEmulator, [host, port]),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<void> applyActionCode(String? code) => (super.noSuchMethod(
-      Invocation.method(#applyActionCode, [code]),
-      returnValue: _i6.Future<void>.value(),
-      returnValueForMissingStub: _i6.Future<void>.value()) as _i6.Future<void>);
-  @override
-  _i6.Future<_i3.ActionCodeInfo> checkActionCode(String? code) =>
-      (super.noSuchMethod(Invocation.method(#checkActionCode, [code]),
-              returnValue: _i6.Future<_i3.ActionCodeInfo>.value(
-                  _FakeActionCodeInfo_1(
-                      this, Invocation.method(#checkActionCode, [code]))))
-          as _i6.Future<_i3.ActionCodeInfo>);
-  @override
-  _i6.Future<void> confirmPasswordReset({String? code, String? newPassword}) =>
-      (super.noSuchMethod(
-              Invocation.method(#confirmPasswordReset, [],
-                  {#code: code, #newPassword: newPassword}),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<_i4.UserCredential> createUserWithEmailAndPassword(
-          {String? email, String? password}) =>
-      (super
-          .noSuchMethod(Invocation.method(#createUserWithEmailAndPassword, [], {#email: email, #password: password}),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#createUserWithEmailAndPassword, [], {
-                        #email: email,
-                        #password: password
-                      })))) as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<List<String>> fetchSignInMethodsForEmail(String? email) => (super
-          .noSuchMethod(Invocation.method(#fetchSignInMethodsForEmail, [email]),
-              returnValue: _i6.Future<List<String>>.value(<String>[]))
-      as _i6.Future<List<String>>);
-  @override
-  _i6.Future<_i4.UserCredential> getRedirectResult() =>
-      (super.noSuchMethod(Invocation.method(#getRedirectResult, []),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#getRedirectResult, []))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  bool isSignInWithEmailLink(String? emailLink) => (super.noSuchMethod(
-      Invocation.method(#isSignInWithEmailLink, [emailLink]),
-      returnValue: false) as bool);
-  @override
-  _i6.Stream<_i4.User?> authStateChanges() =>
-      (super.noSuchMethod(Invocation.method(#authStateChanges, []),
-          returnValue: _i6.Stream<_i4.User?>.empty()) as _i6.Stream<_i4.User?>);
-  @override
-  _i6.Stream<_i4.User?> idTokenChanges() =>
-      (super.noSuchMethod(Invocation.method(#idTokenChanges, []),
-          returnValue: _i6.Stream<_i4.User?>.empty()) as _i6.Stream<_i4.User?>);
-  @override
-  _i6.Stream<_i4.User?> userChanges() =>
-      (super.noSuchMethod(Invocation.method(#userChanges, []),
-          returnValue: _i6.Stream<_i4.User?>.empty()) as _i6.Stream<_i4.User?>);
-  @override
-  _i6.Future<void> sendPasswordResetEmail(
-          {String? email, _i3.ActionCodeSettings? actionCodeSettings}) =>
-      (super.noSuchMethod(
-              Invocation.method(#sendPasswordResetEmail, [],
-                  {#email: email, #actionCodeSettings: actionCodeSettings}),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<void> sendSignInLinkToEmail(
-          {String? email, _i3.ActionCodeSettings? actionCodeSettings}) =>
-      (super.noSuchMethod(
-              Invocation.method(#sendSignInLinkToEmail, [],
-                  {#email: email, #actionCodeSettings: actionCodeSettings}),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<void> setLanguageCode(String? languageCode) => (super.noSuchMethod(
-      Invocation.method(#setLanguageCode, [languageCode]),
-      returnValue: _i6.Future<void>.value(),
-      returnValueForMissingStub: _i6.Future<void>.value()) as _i6.Future<void>);
-  @override
-  _i6.Future<void> setSettings(
-          {bool? appVerificationDisabledForTesting,
-          String? userAccessGroup,
-          String? phoneNumber,
-          String? smsCode,
-          bool? forceRecaptchaFlow}) =>
-      (super.noSuchMethod(
-              Invocation.method(#setSettings, [], {
-                #appVerificationDisabledForTesting:
-                    appVerificationDisabledForTesting,
-                #userAccessGroup: userAccessGroup,
-                #phoneNumber: phoneNumber,
-                #smsCode: smsCode,
-                #forceRecaptchaFlow: forceRecaptchaFlow
-              }),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<void> setPersistence(_i3.Persistence? persistence) =>
-      (super.noSuchMethod(Invocation.method(#setPersistence, [persistence]),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<_i4.UserCredential> signInAnonymously() =>
-      (super.noSuchMethod(Invocation.method(#signInAnonymously, []),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#signInAnonymously, []))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithCredential(
-          _i3.AuthCredential? credential) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithCredential, [credential]),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithCredential, [credential]))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithCustomToken(String? token) =>
-      (super.noSuchMethod(Invocation.method(#signInWithCustomToken, [token]),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithCustomToken, [token]))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithEmailAndPassword(
-          {String? email, String? password}) =>
-      (super
-          .noSuchMethod(Invocation.method(#signInWithEmailAndPassword, [], {#email: email, #password: password}),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#signInWithEmailAndPassword, [], {
-                        #email: email,
-                        #password: password
-                      })))) as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithEmailLink(
-          {String? email, String? emailLink}) =>
-      (super
-          .noSuchMethod(Invocation.method(#signInWithEmailLink, [], {#email: email, #emailLink: emailLink}),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this,
-                      Invocation.method(#signInWithEmailLink, [], {
-                        #email: email,
-                        #emailLink: emailLink
-                      })))) as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithAuthProvider(
-          _i3.AuthProvider? provider) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithAuthProvider, [provider]),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(this,
-                      Invocation.method(#signInWithAuthProvider, [provider]))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<_i4.ConfirmationResult> signInWithPhoneNumber(String? phoneNumber,
-          [_i4.RecaptchaVerifier? verifier]) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInWithPhoneNumber, [phoneNumber, verifier]),
-              returnValue: _i6.Future<_i4.ConfirmationResult>.value(
-                  _FakeConfirmationResult_3(
-                      this,
-                      Invocation.method(
-                          #signInWithPhoneNumber, [phoneNumber, verifier]))))
-          as _i6.Future<_i4.ConfirmationResult>);
-  @override
-  _i6.Future<_i4.UserCredential> signInWithPopup(_i3.AuthProvider? provider) =>
-      (super.noSuchMethod(Invocation.method(#signInWithPopup, [provider]),
-              returnValue: _i6.Future<_i4.UserCredential>.value(
-                  _FakeUserCredential_2(
-                      this, Invocation.method(#signInWithPopup, [provider]))))
-          as _i6.Future<_i4.UserCredential>);
-  @override
-  _i6.Future<void> signInWithRedirect(_i3.AuthProvider? provider) =>
-      (super.noSuchMethod(Invocation.method(#signInWithRedirect, [provider]),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-  @override
-  _i6.Future<void> signOut() => (super.noSuchMethod(
-      Invocation.method(#signOut, []),
-      returnValue: _i6.Future<void>.value(),
-      returnValueForMissingStub: _i6.Future<void>.value()) as _i6.Future<void>);
-  @override
-  _i6.Future<String> verifyPasswordResetCode(String? code) =>
-      (super.noSuchMethod(Invocation.method(#verifyPasswordResetCode, [code]),
-          returnValue: _i6.Future<String>.value('')) as _i6.Future<String>);
-  @override
-  _i6.Future<void> verifyPhoneNumber(
-          {String? phoneNumber,
-          _i3.PhoneMultiFactorInfo? multiFactorInfo,
-          _i3.PhoneVerificationCompleted? verificationCompleted,
-          _i3.PhoneVerificationFailed? verificationFailed,
-          _i3.PhoneCodeSent? codeSent,
-          _i3.PhoneCodeAutoRetrievalTimeout? codeAutoRetrievalTimeout,
-          String? autoRetrievedSmsCodeForTesting,
-          Duration? timeout = const Duration(seconds: 30),
-          int? forceResendingToken,
-          _i3.MultiFactorSession? multiFactorSession}) =>
-      (super.noSuchMethod(
-              Invocation.method(#verifyPhoneNumber, [], {
-                #phoneNumber: phoneNumber,
-                #multiFactorInfo: multiFactorInfo,
-                #verificationCompleted: verificationCompleted,
-                #verificationFailed: verificationFailed,
-                #codeSent: codeSent,
-                #codeAutoRetrievalTimeout: codeAutoRetrievalTimeout,
-                #autoRetrievedSmsCodeForTesting: autoRetrievedSmsCodeForTesting,
-                #timeout: timeout,
-                #forceResendingToken: forceResendingToken,
-                #multiFactorSession: multiFactorSession
-              }),
-              returnValue: _i6.Future<void>.value(),
-              returnValueForMissingStub: _i6.Future<void>.value())
-          as _i6.Future<void>);
-}
-
-/// A class which mocks [GoogleSignIn].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockGoogleSignIn extends _i1.Mock implements _i5.GoogleSignIn {
-  MockGoogleSignIn() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i7.SignInOption get signInOption =>
-      (super.noSuchMethod(Invocation.getter(#signInOption),
-          returnValue: _i7.SignInOption.standard) as _i7.SignInOption);
-  @override
-  List<String> get scopes =>
-      (super.noSuchMethod(Invocation.getter(#scopes), returnValue: <String>[])
-          as List<String>);
-  @override
-  _i6.Stream<_i5.GoogleSignInAccount?> get onCurrentUserChanged =>
-      (super.noSuchMethod(Invocation.getter(#onCurrentUserChanged),
-              returnValue: _i6.Stream<_i5.GoogleSignInAccount?>.empty())
-          as _i6.Stream<_i5.GoogleSignInAccount?>);
-  @override
-  _i6.Future<_i5.GoogleSignInAccount?> signInSilently(
-          {bool? suppressErrors = true, bool? reAuthenticate = false}) =>
-      (super.noSuchMethod(
-              Invocation.method(#signInSilently, [], {
-                #suppressErrors: suppressErrors,
-                #reAuthenticate: reAuthenticate
-              }),
-              returnValue: _i6.Future<_i5.GoogleSignInAccount?>.value())
-          as _i6.Future<_i5.GoogleSignInAccount?>);
-  @override
-  _i6.Future<bool> isSignedIn() =>
-      (super.noSuchMethod(Invocation.method(#isSignedIn, []),
-          returnValue: _i6.Future<bool>.value(false)) as _i6.Future<bool>);
-  @override
-  _i6.Future<_i5.GoogleSignInAccount?> signIn() =>
-      (super.noSuchMethod(Invocation.method(#signIn, []),
-              returnValue: _i6.Future<_i5.GoogleSignInAccount?>.value())
-          as _i6.Future<_i5.GoogleSignInAccount?>);
-  @override
-  _i6.Future<_i5.GoogleSignInAccount?> signOut() =>
-      (super.noSuchMethod(Invocation.method(#signOut, []),
-              returnValue: _i6.Future<_i5.GoogleSignInAccount?>.value())
-          as _i6.Future<_i5.GoogleSignInAccount?>);
-  @override
-  _i6.Future<_i5.GoogleSignInAccount?> disconnect() =>
-      (super.noSuchMethod(Invocation.method(#disconnect, []),
-              returnValue: _i6.Future<_i5.GoogleSignInAccount?>.value())
-          as _i6.Future<_i5.GoogleSignInAccount?>);
-  @override
-  _i6.Future<bool> requestScopes(List<String>? scopes) =>
-      (super.noSuchMethod(Invocation.method(#requestScopes, [scopes]),
-          returnValue: _i6.Future<bool>.value(false)) as _i6.Future<bool>);
-}
-
-/// A class which mocks [GoogleSignInAccount].
-///
-/// See the documentation for Mockito's code generation for more information.
-// ignore: must_be_immutable
-class MockGoogleSignInAccount extends _i1.Mock
-    implements _i5.GoogleSignInAccount {
-  MockGoogleSignInAccount() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  String get email =>
-      (super.noSuchMethod(Invocation.getter(#email), returnValue: '')
-          as String);
-  @override
-  String get id =>
-      (super.noSuchMethod(Invocation.getter(#id), returnValue: '') as String);
-  @override
-  _i6.Future<_i5.GoogleSignInAuthentication> get authentication =>
-      (super.noSuchMethod(Invocation.getter(#authentication),
-              returnValue: _i6.Future<_i5.GoogleSignInAuthentication>.value(
-                  _FakeGoogleSignInAuthentication_4(
-                      this, Invocation.getter(#authentication))))
-          as _i6.Future<_i5.GoogleSignInAuthentication>);
-  @override
-  _i6.Future<Map<String, String>> get authHeaders =>
-      (super.noSuchMethod(Invocation.getter(#authHeaders),
-              returnValue:
-                  _i6.Future<Map<String, String>>.value(<String, String>{}))
-          as _i6.Future<Map<String, String>>);
-  @override
-  _i6.Future<void> clearAuthCache() => (super.noSuchMethod(
-      Invocation.method(#clearAuthCache, []),
-      returnValue: _i6.Future<void>.value(),
-      returnValueForMissingStub: _i6.Future<void>.value()) as _i6.Future<void>);
-}
-
-/// A class which mocks [GoogleSignInAuthentication].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockGoogleSignInAuthentication extends _i1.Mock
-    implements _i5.GoogleSignInAuthentication {
-  MockGoogleSignInAuthentication() {
-    _i1.throwOnMissingStub(this);
-  }
-}
-
-/// A class which mocks [UserCredential].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockUserCredential extends _i1.Mock implements _i4.UserCredential {
-  MockUserCredential() {
-    _i1.throwOnMissingStub(this);
-  }
-}
